# docs/swagger.yaml
basePath: /
definitions:
...
  models.CreateTemporadaDTO:
    properties:
      fecha_fin:
        type: string
      fecha_inicio:
        type: string
      id_estado_temporada:
        type: integer
      nombre_temporada:
        type: string
    required:
    - fecha_fin
    - fecha_inicio
    - nombre_temporada
    type: object
...
  models.ErrorInfo:
    properties:
      code:
        type: string
      details: {}
    type: object
...
  models.StandardResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/models.ErrorInfo'
      errors: {}
      message:
        type: string
      meta: {}
      success:
        type: boolean
    type: object
...
host: localhost:8084
info:
  contact:
    email: soporte@sistal.com
    name: Soporte API
  description: |-
    API del Microservicio de Catálogo de Prendas y Uniformes
    Gestiona el catálogo de prendas, uniformes, temporadas y datos maestros
  license:
    name: MIT
  title: MS-Catalogo API
  version: "1.0"
paths:
...
  /api/v1/temporadas:
    get:
      consumes:
      - application/json
      description: Obtiene lista de temporadas/campañas con paginación
      parameters:
      - default: 1
        description: Número de página
        in: query
        name: page
        type: integer
      - default: 20
        description: Límite por página
        in: query
        name: limit
        type: integer
      - description: Filtrar por estado (40=Activa, 41=Inactiva, 42=Finalizada)
        in: query
        name: id_estado_temporada
        type: integer
      - description: Solo temporada activa
        in: query
        name: activa_solo
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StandardResponse'
      summary: Listar Temporadas
      tags:
      - Temporadas
    post:
      consumes:
      - application/json
      description: Crea una nueva temporada
      parameters:
      - description: Datos de la temporada
        in: body
        name: temporada
        required: true
        schema:
          $ref: '#/definitions/models.CreateTemporadaDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardResponse'
      summary: Crear Temporada
      tags:
      - Temporadas
  /api/v1/temporadas/{id_temporada}/activar:
    patch:
      consumes:
      - application/json
      description: Activa una temporada (desactiva automáticamente la temporada actual)
      parameters:
      - description: ID de la temporada
        in: path
        name: id_temporada
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.StandardResponse'
      summary: Activar Temporada
      tags:
      - Temporadas
  /api/v1/temporadas/activa:
    get:
      consumes:
      - application/json
      description: Obtiene la temporada actualmente activa
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.StandardResponse'
      summary: Obtener Temporada Activa
      tags:
      - Temporadas
...
schemes:
- http
- https
swagger: "2.0"
...